Add participant magic guard

+++ src/agenda-items/agenda-items.controller.ts
@@ -1,8 +1,11 @@
+import { MeetingsService } from './../meetings/meetings.service';
 import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
+  NotFoundException,
   Param,
   ParseIntPipe,
   ParseUUIDPipe,
@@ -16,13 +19,18 @@ import {
   ApiParam,
   ApiTags,
 } from '@nestjs/swagger';
-import { UseBearerAuth } from '../shared/decorators/auth.decorator';
+import { UseAuth, UseBearerAuth } from '../shared/decorators/auth.decorator';
 import { AgendaItem } from './agenda-item.entity';
 import { AgendaItemsService } from './agenda-items.service';
 import { CreateAgendaItemDto } from './dto/create-agenda-item.dto';
 import { UpdateAgendaItemDto } from './dto/update-agenda-item.dto';
 import { UpdateAgendaItemsPositionDto } from './dto/update-agenda-items-position.dto';
 import { MeetingSocketGateway } from '../meeting-socket/meeting-socket.gateway';
+import { AccessGuard } from 'src/participants/guard/access.guard';
+import { Usr } from 'src/shared/decorators/user.decorator';
+import { User } from 'src/users/user.entity';
+import { Participant } from 'src/participants/participant.entity';
+import { AccessUser } from 'src/shared/decorators/participant.decorator';
 
 @ApiTags('Agenda Item')
 @Controller('agenda-item')
@@ -30,6 +38,7 @@ export class AgendaItemsController {
   constructor(
     private readonly agendaItemsService: AgendaItemsService,
     private readonly meetingGateway: MeetingSocketGateway,
+    private readonly meetingsService: MeetingsService,
   ) {}
 
   @ApiCreatedResponse({
@@ -40,8 +49,17 @@ export class AgendaItemsController {
   @ApiBody({ type: CreateAgendaItemDto })
   @Post('/')
   public async createAgendaItems(
+    @Usr() user: User,
     @Body() createAgendaItemDto: CreateAgendaItemDto,
   ): Promise<AgendaItem> {
+    if (
+      !(await this.meetingsService.isHostOfMeeting(
+        user.uuid,
+        createAgendaItemDto.meetingId,
+      ))
+    ) {
+      throw new ForbiddenException('Not allowed to create');
+    }
     return this.agendaItemsService.createOneAgendaItem(createAgendaItemDto);
   }
 
@@ -53,11 +71,27 @@ export class AgendaItemsController {
     name: 'meetingUuid',
     description: 'The id of the meeting',
   })
-  @UseBearerAuth()
+  @UseAuth(AccessGuard)
   @Get('/:meetingUuid')
   public async getAgendaItemsByMeetingId(
+    @AccessUser() userOrParticipant: User | Participant,
     @Param('meetingUuid', ParseUUIDPipe) meetingId: string,
   ): Promise<AgendaItem[]> {
+    if (
+      userOrParticipant['meetingId'] &&
+      (userOrParticipant as Participant).meetingId !== meetingId
+    ) {
+      throw new ForbiddenException('Not allowed to access meeting');
+    } else if (
+      userOrParticipant['uuid'] &&
+      !(await this.meetingsService.isHostOfMeeting(
+        (userOrParticipant as User).uuid,
+        meetingId,
+      ))
+    ) {
+      throw new ForbiddenException('Not allowed to access meeting');
+    }
+
     return this.agendaItemsService.getAgendaItemsByMeetingId(meetingId);
   }
 
@@ -72,9 +106,15 @@ export class AgendaItemsController {
   @UseBearerAuth()
   @Delete('/:meetingUuid/:position')
   public async deleteAgendaItemByPosition(
+    @Usr() requester: User,
     @Param('meetingUuid', ParseUUIDPipe) meetingId: string,
     @Param('position', ParseIntPipe) position: number,
   ): Promise<void> {
+    if (
+      !(await this.meetingsService.isHostOfMeeting(requester.uuid, meetingId))
+    ) {
+      throw new ForbiddenException('Not allowed to delete agenda');
+    }
     await this.agendaItemsService.deleteAgendaItemByMeetingIdAndPosition(
       meetingId,
       position,
@@ -88,16 +128,38 @@ export class AgendaItemsController {
   @ApiBody({
     type: UpdateAgendaItemDto,
   })
-  // @UseBearerAuth()
+  @UseAuth(AccessGuard)
   @Put('/:meetingUuid/:position')
   public async updateAgendaItemByPosition(
+    @AccessUser() userOrParticipant: User | Participant,
     @Param('meetingUuid', ParseUUIDPipe) meetingId: string,
     @Param('position', ParseIntPipe) position: number,
     @Body() updateAgendaItemDto: UpdateAgendaItemDto,
   ): Promise<void> {
+    const targetAgenda =
+      await this.agendaItemsService.getAgendaItemByMeetingIdAndPosition(
+        meetingId,
+        position,
+      );
+    if (!targetAgenda) {
+      throw new NotFoundException(`Agenda item not found`);
+    }
+
+    if (
+      userOrParticipant['meetingId'] &&
+      ((userOrParticipant as Participant).meetingId !== meetingId ||
+        (userOrParticipant as Participant).id !== targetAgenda?.speaker?.id)
+    ) {
+      throw new ForbiddenException('Not allowed to update agenda');
+    } else if (
+      userOrParticipant['uuid'] &&
+      (userOrParticipant as User).uuid !== targetAgenda?.meeting?.hostId
+    ) {
+      throw new ForbiddenException('Not allowed to access meeting');
+    }
+
     await this.agendaItemsService.updateAgendaItemByMeetingIdAndPosition(
-      meetingId,
-      position,
+      targetAgenda,
       updateAgendaItemDto,
     );
     this.meetingGateway.emitAgendaUpdated(meetingId);
@@ -112,8 +174,17 @@ export class AgendaItemsController {
   @UseBearerAuth()
   @Put('/positions')
   public async reorderAgendaItemsPosition(
+    @Usr() requester: User,
     @Body() updateAgendaItemsPositionDto: UpdateAgendaItemsPositionDto,
   ): Promise<void> {
+    if (
+      !(await this.meetingsService.isHostOfMeeting(
+        requester.uuid,
+        updateAgendaItemsPositionDto.meetingId,
+      ))
+    ) {
+      throw new ForbiddenException('Not allowed to move agenda');
+    }
     await this.agendaItemsService.reorderAgendaItemsPosition(
       updateAgendaItemsPositionDto,
     );
